+++ R/write_problem.R
@@ -1,0 +1,87 @@
+#' @include internal.R
+NULL
+
+#' Write problem
+#'
+#' Save the mathematical formulation for a conservation planning [problem()]
+#' to a file for mixed integer programming solvers. Note that this function
+#' requires the \pkg{Rsymphony} package to be installed.
+#'
+#' @inheritParams compile
+#'
+#' @param path `character` file path to save the problem formulation.
+#' The argument should contain a `".lp"` or `.mps"` file extension
+#' to specify whether the problem formulation will be saved in the
+#' [LP](https://www.gurobi.com/documentation/9.1/refman/lp_format.html) or
+#' [MPS](https://www.gurobi.com/documentation/9.1/refman/mps_format.html)
+#' format (respectively).
+#'
+#' @return Invisible `TRUE` indicating success.
+#'
+#' @examples
+#' \dontrun{
+#' # set seed for reproducibility
+#' set.seed(500)
+#'
+#' # load data
+#' data(sim_pu_raster, sim_features, sim_pu_zones_stack, sim_features_zones)
+#'
+#' # create minimal problem
+#' p <- problem(sim_pu_raster, sim_features) %>%
+#'       add_min_set_objective() %>%
+#'       add_relative_targets(0.1) %>%
+#'       add_binary_decisions()
+#'
+#' # save problem to file (using the Rsymphony R package)
+#' write_problem(p, "model.lp")
+#' }
+#'
+#' @export
+write_problem <- function(x, path) {
+  # assert arguments are valid
+  assertthat::assert_that(
+    inherits(x, "ConservationProblem"),
+    assertthat::is.string(path),
+    assertthat::noNA(path),
+    requireNamespace("Rsymphony"),
+    assertthat:: is.writeable(dirname(path)))
+  assertthat::assert_that(
+    assertthat::has_extension(path, "lp") ||
+    assertthat::has_extension(path, "mps"),
+    msg = "argument to path must have a \".lp\" or \".mps\" file extension")
+
+  # add rsymphony solver
+  x <- suppressWarnings(add_rsymphony_solver(x))
+
+  # compile problem
+  x$solver$calculate(compile(x))
+  model <- x$solver$data$model
+  assertthat::assert_that(
+    is.list(model),
+    msg = "failed to compile optimization problem")
+
+  # create parameters
+  p <- list(
+    first_feasible = TRUE,
+    verbosity = -2,
+    node_limit = 1,
+    write_lp = endsWith(path, ".lp"),
+    write_mps = endsWith(path, ".mps"))
+
+  # create temporary directory to save result
+  d <- file.path(tempdir(), basename(tempfile()))
+  dir.create(d, showWarnings = FALSE, recursive = TRUE)
+
+  # save problem to disk by generating a solution to the problem
+  y <- withr::with_dir(
+    d, do.call(Rsymphony::Rsymphony_solve_LP, append(model, p)))
+
+  # move file to desired location
+  assertthat::assert_that(
+    length(dir(d)) == 1,
+    msg = "failed to save problem formulation")
+  file.rename(dir(d, full.names = TRUE)[1], path)
+
+  # return success
+  invisible(TRUE)
+}